+++ corehq/motech/openmrs/repeaters.py
@@ -17,11 +17,11 @@ from corehq.motech.openmrs.const import XMLNS_OPENMRS
 from corehq.motech.openmrs.openmrs_config import OpenmrsConfig
 from corehq.motech.openmrs.handler import send_openmrs_data
 from corehq.motech.openmrs.repeater_helpers import (
-    get_form_question_values,
     get_relevant_case_updates_from_form_json,
     get_case_location_ancestor_repeaters,
 )
 from corehq.motech.requests import Requests
+from corehq.motech.value_source import get_form_question_values
 from memoized import memoized
