+++ corehq/motech/value_source.py
@@ -116,3 +116,39 @@ class FormQuestionMap(FormQuestion):
             return self.value_map[value]
         except KeyError:
             return None
+
+
+def get_form_question_values(form_json):
+    """
+    Returns question-value pairs to result where questions are given as "/data/foo/bar"
+
+    >>> values = get_form_question_values({'form': {'foo': {'bar': 'baz'}}})
+    >>> values == {'/data/foo/bar': 'baz'}
+    True
+
+    """
+    _reserved_keys = ('@uiVersion', '@xmlns', '@name', '#type', 'case', 'meta', '@version')
+
+    def _recurse_form_questions(form_dict, path, result_):
+        for key, value in form_dict.items():
+            if key in _reserved_keys:
+                continue
+            new_path = path + [key]
+            if isinstance(value, list):
+                # Repeat group
+                for v in value:
+                    assert isinstance(v, dict)
+                    _recurse_form_questions(v, new_path, result_)
+            elif isinstance(value, dict):
+                # Group
+                _recurse_form_questions(value, new_path, result_)
+            else:
+                # key is a question and value is its answer
+                question = '/'.join((p.decode('utf8') if isinstance(p, bytes) else p for p in new_path))
+                result_[question] = value
+
+    result = {}
+    _recurse_form_questions(form_json['form'], [b'/data'], result)  # "/data" is just convention, hopefully
+    # familiar from form builder. The form's data will usually be immediately under "form_json['form']" but not
+    # necessarily. If this causes problems we may need a more reliable way to get to it.
+    return result
